// actions/supadmin/Create-Student_school_Timetable/schema.ts
import { z } from "zod";

export const Supstudent_school_timetable_Create_Schema = z.object({
    supadminId: z.string(),
    name: z.string(),
    img: z.string(),
    student_school_timetable_id: z.string(),
    student_name: z.string(),
    grade: z.number(),
    year: z.string(),
    quarter: z.number(),
    school: z.string(),
    parentId: z.string(),
})